fix delete bugs

+++ src/input.h
@@ -10,5 +10,5 @@ void editorProcessKeyPress(void);
 int getCursorPosition(int* rows, int* cols);
 void abAppend(struct abuf* ab, const char* s, int len);
 void abFree(struct abuf* ab);
-void rowAppendString(rowOfText* row, char* s, int len);
+void rowAppendString(rowOfText* row, char* s, size_t len);
 #endif
